bertopic model building into scripts and nlp modules and writing to generated_data dir, with other generated_data path tweaks
