+++ app/app_metadata.py
@@ -1,4 +1,9 @@
-PATH_TO_GENERATED_DATA = './generated_data'
+GENERATED_DATA_DIR = './generated_data'
 
-PATH_TO_SENTIMENT_DATA = f'{PATH_TO_GENERATED_DATA}/sentiment'
-PATH_TO_ANALYTICS_DATA = f'{PATH_TO_GENERATED_DATA}/analytics'
+SENTIMENT_DATA_DIR = f'{GENERATED_DATA_DIR}/sentiment'
+ANALYTICS_DIR = f'{GENERATED_DATA_DIR}/analytics'
+BERTOPIC_DIR = f'{GENERATED_DATA_DIR}/bertopic'
+
+BERTOPIC_DATA_DIR = f'{BERTOPIC_DIR}/data'
+BERTOPIC_MODELS_DIR = f'{BERTOPIC_DIR}/models'
+BERTOPIC_LOGS_DIR = f'{BERTOPIC_DIR}/logs'
