+++ app/nlp/bertopic_model_builder.py
@@ -1,167 +1,25 @@
-import argparse
 from bertopic import BERTopic
 from bertopic.representation import KeyBERTInspired, MaximalMarginalRelevance, OpenAI as BertOpenAI
-import datetime
 from hdbscan import HDBSCAN
-from itertools import chain, product
-# from huggingface_hub import login
+from itertools import chain
 import math
 from operator import itemgetter
 import os
 import openai
-# from openai import OpenAI
 import pandas as pd
-import pickle
-# import plotly.express as px
 from sentence_transformers import SentenceTransformer
 from sklearn.feature_extraction.text import CountVectorizer
 from umap import UMAP
 
+from app.app_metadata import BERTOPIC_DATA_DIR, BERTOPIC_MODELS_DIR, BERTOPIC_LOGS_DIR
 from app.config import settings
 import app.es.es_read_router as esr
-from app.nlp.nlp_metadata import MIN_WORDS_FOR_BERT, MAX_WORDS_FOR_BERT, BERTOPIC_DATA_DIR, BERTOPIC_MODELS_DIR
+from app.nlp.nlp_metadata import MIN_WORDS_FOR_BERT, MAX_WORDS_FOR_BERT
 from app.show_metadata import ShowKey
 import app.utils as utils
 
 
 openai_client = openai.OpenAI(api_key=settings.openai_api_key)
-# openai.api_key = settings.openai_api_key
-
-LOG_FILE = 'bertopic_config_tests/bertopic_config_test'
-REPORT_FILE = 'bertopic_config_tests/bertopic_config_report'
-
-
-CONFIG_OPTIONS = {
-    # 'narrative_only': [True, False],
-    'sentence_transformer_lm': 'all-MiniLM-L12-v2',
-    'vec_ngram_low': 1,
-    'vec_ngram_high': 2,
-    'bertopic_top_n_words': 5,
-    'umap_n_neighbors': 5,
-    'umap_n_components': 3,
-    'umap_min_dist': [0, 0.01, 0.02],
-    # 'invalid_umap_metric': ['euclidian', 'seuclidian', 'haversine', 'mahalanobis'],
-    # 'unusable_umap_metric': ['wminkowski'],  # wminkowski: best correlation to vector-based topics, but doesn't translate to 3D clusters
-    # 'umap_metric': ['braycurtis', 'minkowski', 'canberra', 'manhattan', 'cosine', 'correlation'],
-    'umap_metric': ['braycurtis', 'minkowski', 'canberra', 'manhattan', 'cosine', 'correlation'],
-    'umap_random_state': [4, 53, 87],
-    'hdbscan_min_cluster_size': [25, 50],
-    'hdbscan_min_samples': 10,
-    # 'mmr_diversity': [0.05, 0.1, 0.15, 0.2, 0.3, 0.4, 0.5],
-    'mmr_diversity': 0.05,
-    # 'representation_model_type': ['Custom', 'MaximalMarginalRelevance', 'KeyBERTInspired', 'BertOpenAI'],
-    'topic_count_threshold': 6,
-    'topic_ratio_threshold': 0.5,
-    'corr_threshold': 0.4,
-    'narrative_freq_threshold': 7,
-    'match_threshold': 1,
-}
-
-
-def main():
-    ts_log = str(datetime.datetime.now())[:19]
-    ts_filename = ts_log.replace(' ', '_').replace('-', '').replace(':', '')
-    print(f'begin bertopic_modeling at {ts_log}')
-    # parse script params
-    parser = argparse.ArgumentParser()
-    parser.add_argument("--show_key", "-s", help="Show key", required=True)
-    parser.add_argument("--umap_metric", "-m", help="UMAP metric", required=False)
-    parser.add_argument("--umap_random_state", "-r", help="UMAP random state", required=False)
-    parser.add_argument("--umap_min_dist", "-d", help="UMAP min dist", required=False)
-    parser.add_argument("--hdbscan_min_cluster_size", "-c", help="HDBSCAN min cluster size", required=False)
-    args = parser.parse_args()
-    # assign script params to vars
-    show_key = args.show_key
-    override_config = {}
-    if args.umap_metric:
-        override_config['umap_metric'] = args.umap_metric.split(',')
-    if args.umap_random_state:
-        override_config['umap_random_state'] = [int(v) for v in args.umap_random_state.split(',')]
-    if args.umap_min_dist:
-        override_config['umap_min_dist'] = [float(v) for v in args.umap_min_dist.split(',')]
-    if args.hdbscan_min_cluster_size:
-        override_config['hdbscan_min_cluster_size'] = [int(v) for v in args.hdbscan_min_cluster_size.split(',')]
-
-    # print(f'attempt huggingface login')
-    # login()
-    # print(f'huggingface login success')
-        
-    print(f"override_config={override_config}")
-    configs, config_params_to_values = generate_configs(override_config)
-
-    bert_text_inputs, bert_text_sources = generate_bert_text_inputs(ShowKey(show_key), narrative_only=False)
-
-    simple_episodes_response = esr.fetch_simple_episodes(ShowKey(show_key))
-    episodes = simple_episodes_response['episodes']
-    episodes_df = pd.DataFrame(episodes)
-    sources_df = pd.DataFrame(bert_text_sources, columns=['episode_key', 'speaker_group', 'wc'])
-    sources_df = pd.merge(sources_df, episodes_df, on='episode_key')
-
-    log_file = open(f'{LOG_FILE}_{ts_filename}.csv', 'a')  # append mode
-    log_file.write('=======================================================================================\n')
-    log_file.write(f'Begin new job for {len(configs)} config variations at {ts_log}\n')
-    log_file.write(f'config_params_to_values: {config_params_to_values}\n')
-    log_file.close()
-
-    report_dicts = []
-    i = 1
-    for config in configs:
-        print(f"processing config {i} of {len(configs)}")
-        report_dict = generate_bertopic_models(show_key, bert_text_inputs, config, sources_df)
-        i += 1
-        if report_dict:
-            report_dicts.append(report_dict)
-
-    # log summarized output to dataframe csv
-    if report_dicts:
-        report_df = pd.DataFrame(report_dicts)
-        report_file_name = f'{REPORT_FILE}_{ts_filename}.csv'
-        report_df.to_csv(report_file_name, sep='\t')
-
-
-def generate_configs(config_overrides: dict = None) -> tuple[list, dict]:
-    # incorporate overrides
-    config_options = CONFIG_OPTIONS
-    if config_overrides:
-        for override_param, override_values in config_overrides.items():
-            config_options[override_param] = override_values
-    # generate 'flattened' single-value config variants out of multi-value config
-    config_params = []
-    config_params_to_values = {}
-    config_param_value_counts = []
-    for config_param, config_values in config_options.items():
-        if isinstance(config_values, list):
-            config_params.append(config_param)
-            config_params_to_values[config_param] = config_values
-            config_param_value_counts.append(len(config_values))
-    # print(f'config_params={config_params}')
-    # print(f'config_params_to_values={config_params_to_values}')
-    # print(f'config_param_value_counts={config_param_value_counts}')
-
-    mx = []
-    for vc in config_param_value_counts:
-        mx.append([i for i in range(vc)])
-    config_combos = list(product(*mx))
-    # print(f'config_combos={config_combos}')
-
-    configs = []
-    for config_combo in config_combos:
-        config = generate_config_instance(config_params, config_combo)
-        configs.append(config)
-
-    return configs, config_params_to_values
-
-
-def generate_config_instance(config_params: list, config_value_indexes: tuple) -> dict:
-    config_instance = {}
-    for config_param, config_val in CONFIG_OPTIONS.items():
-        if isinstance(config_val, list):
-            config_param_i = config_params.index(config_param)
-            config_val_i = config_value_indexes[config_param_i]
-            config_instance[config_param] = config_val[config_val_i]
-        else:
-            config_instance[config_param] = config_val
-    return config_instance
 
 
 def generate_bertopic_models(show_key: str, bert_text_inputs: list, config: dict, sources_df: pd.DataFrame) -> dict|None:
@@ -326,17 +184,17 @@ def generate_bertopic_models(show_key: str, bert_text_inputs: list, config: dict
     bertopic_docs_df = pd.get_dummies(bertopic_docs_df, columns=['Topic'], dtype=int, drop_first=False)
     
     # vector topics and speaker_group: convert to list, then one-hot encode values from the resulting 'list' column 
-    topics_focused_prefix = 'foctopic_'
-    topics_focused_tfidf_prefix = 'foctopictfidf_'
+    # topics_focused_prefix = 'foctopic_'
+    # topics_focused_tfidf_prefix = 'foctopictfidf_'
     topics_universal_prefix = 'univtopic_'
     topics_universal_tfidf_prefix = 'univtopictfidf_'
     speaker_prefix = 'spkr_'
-    # topics_focused
-    bertopic_docs_df['topics_focused_list'] = bertopic_docs_df['topics_focused'].apply(lambda x: [t['topic_key'] for t in x[:3]])
-    bertopic_docs_df = one_hot_multival_col(bertopic_docs_df, topics_focused_prefix, 'topics_focused_list')
-    # topics_focused_tfidf
-    bertopic_docs_df['topics_focused_tfidf_list'] = bertopic_docs_df['topics_focused_tfidf'].apply(lambda x: [t['topic_key'] for t in x[:3]])
-    bertopic_docs_df = one_hot_multival_col(bertopic_docs_df, topics_focused_tfidf_prefix, 'topics_focused_tfidf_list')
+    # # topics_focused
+    # bertopic_docs_df['topics_focused_list'] = bertopic_docs_df['topics_focused'].apply(lambda x: [t['topic_key'] for t in x[:3]])
+    # bertopic_docs_df = one_hot_multival_col(bertopic_docs_df, topics_focused_prefix, 'topics_focused_list')
+    # # topics_focused_tfidf
+    # bertopic_docs_df['topics_focused_tfidf_list'] = bertopic_docs_df['topics_focused_tfidf'].apply(lambda x: [t['topic_key'] for t in x[:3]])
+    # bertopic_docs_df = one_hot_multival_col(bertopic_docs_df, topics_focused_tfidf_prefix, 'topics_focused_tfidf_list')
     # topics_universal
     bertopic_docs_df['topics_universal_list'] = bertopic_docs_df['topics_universal'].apply(lambda x: [t['topic_key'] for t in x[:3]])
     bertopic_docs_df = one_hot_multival_col(bertopic_docs_df, topics_universal_prefix, 'topics_universal_list')
@@ -368,28 +226,28 @@ def generate_bertopic_models(show_key: str, bert_text_inputs: list, config: dict
     #### RESULTS ANALYSIS PT 2 ####
     # generate corr dfs
     bertopic_cols = [c for c in bertopic_docs_df.columns if bertopic_prefix in c]
-    foctopic_cols = [c for c in bertopic_docs_df.columns if topics_focused_prefix in c]
-    foctopictfidf_cols = [c for c in bertopic_docs_df.columns if topics_focused_tfidf_prefix in c]
+    # foctopic_cols = [c for c in bertopic_docs_df.columns if topics_focused_prefix in c]
+    # foctopictfidf_cols = [c for c in bertopic_docs_df.columns if topics_focused_tfidf_prefix in c]
     univtopic_cols = [c for c in bertopic_docs_df.columns if topics_universal_prefix in c]
     univtopictfidf_cols = [c for c in bertopic_docs_df.columns if topics_universal_tfidf_prefix in c]
     speaker_cols = [c for c in bertopic_docs_df.columns if speaker_prefix in c]
 
     # bertopic_cols corrs to vector topic and speaker cols
     # save_models = False
-    corr_values_foctopic_x_bertopic = extract_corr_values(bertopic_docs_df, foctopic_cols, bertopic_cols, corr_threshold, narrative_freq_threshold)
-    if len(corr_values_foctopic_x_bertopic) >= match_threshold:
-        logs.append(f'corr_values_foctopic_x_bertopic={corr_values_foctopic_x_bertopic}')
-        report_dict['foctopic_x_bertopic'] = len(corr_values_foctopic_x_bertopic)
-        # save_models = True
-    else:
-        report_dict['foctopic_x_bertopic'] = 0
-    corr_values_foctopictfidf_x_bertopic = extract_corr_values(bertopic_docs_df, foctopictfidf_cols, bertopic_cols, corr_threshold, narrative_freq_threshold)
-    if len(corr_values_foctopictfidf_x_bertopic) >= match_threshold:
-        logs.append(f'corr_values_foctopictfidf_x_bertopic={corr_values_foctopictfidf_x_bertopic}')
-        report_dict['foctopictfidf_x_bertopic'] = len(corr_values_foctopictfidf_x_bertopic)
-        # save_models = True
-    else:
-        report_dict['foctopictfidf_x_bertopic'] = 0
+    # corr_values_foctopic_x_bertopic = extract_corr_values(bertopic_docs_df, foctopic_cols, bertopic_cols, corr_threshold, narrative_freq_threshold)
+    # if len(corr_values_foctopic_x_bertopic) >= match_threshold:
+    #     logs.append(f'corr_values_foctopic_x_bertopic={corr_values_foctopic_x_bertopic}')
+    #     report_dict['foctopic_x_bertopic'] = len(corr_values_foctopic_x_bertopic)
+    #     # save_models = True
+    # else:
+    #     report_dict['foctopic_x_bertopic'] = 0
+    # corr_values_foctopictfidf_x_bertopic = extract_corr_values(bertopic_docs_df, foctopictfidf_cols, bertopic_cols, corr_threshold, narrative_freq_threshold)
+    # if len(corr_values_foctopictfidf_x_bertopic) >= match_threshold:
+    #     logs.append(f'corr_values_foctopictfidf_x_bertopic={corr_values_foctopictfidf_x_bertopic}')
+    #     report_dict['foctopictfidf_x_bertopic'] = len(corr_values_foctopictfidf_x_bertopic)
+    #     # save_models = True
+    # else:
+    #     report_dict['foctopictfidf_x_bertopic'] = 0
     corr_values_univtopic_x_bertopic = extract_corr_values(bertopic_docs_df, univtopic_cols, bertopic_cols, corr_threshold, narrative_freq_threshold)
     if len(corr_values_univtopic_x_bertopic) >= match_threshold:
         logs.append(f'corr_values_univtopic_x_bertopic={corr_values_univtopic_x_bertopic}')
@@ -416,7 +274,7 @@ def generate_bertopic_models(show_key: str, bert_text_inputs: list, config: dict
     print('-----------------------------------------------------------------------------------------')
     print(f'config={config}')
     print(f'logs={logs}')
-    f = open(LOG_FILE, 'a')  # append mode
+    f = open(f'{BERTOPIC_LOGS_DIR}/log_{show_key}', 'a')  # append mode
     f.write('-----------------------------------------------------------------------------------------\n')
     f.write(f'config: {config}\n')
     for log in logs:
@@ -427,19 +285,19 @@ def generate_bertopic_models(show_key: str, bert_text_inputs: list, config: dict
     # if save_models:
     model_id = f'{umap_metric}_{umap_random_state}_{umap_min_dist}_{hdbscan_min_cluster_size}'
     # create model dirs
-    bertopic_models_dir = f'{BERTOPIC_MODELS_DIR}/{show_key}'
-    bertopic_data_dir = f'{BERTOPIC_DATA_DIR}/{show_key}'
-    os.mkdir(f"{bertopic_models_dir}/{model_id}")
-    os.mkdir(f"{bertopic_models_dir}/{model_id}/mmr")
-    os.mkdir(f"{bertopic_models_dir}/{model_id}/kb")
-    os.mkdir(f"{bertopic_models_dir}/{model_id}/openai")
+    # bertopic_models_dir = f'{BERTOPIC_MODELS_DIR}/{show_key}'
+    # bertopic_data_dir = f'{BERTOPIC_DATA_DIR}/{show_key}'
+    os.mkdir(f'{BERTOPIC_MODELS_DIR}/{show_key}/{model_id}')
+    os.mkdir(f'{BERTOPIC_MODELS_DIR}/{show_key}/{model_id}/mmr')
+    os.mkdir(f'{BERTOPIC_MODELS_DIR}/{show_key}/{model_id}/kb')
+    os.mkdir(f'{BERTOPIC_MODELS_DIR}/{show_key}/{model_id}/openai')
     # save models
-    se_model = f"sentence-transformers/{sentence_transformer_lm}"
-    mmr_bertopic_model.save(f"{bertopic_models_dir}/{model_id}/mmr", serialization="safetensors", save_ctfidf=True, save_embedding_model=se_model)
-    kb_bertopic_model.save(f"{bertopic_models_dir}/{model_id}/kb", serialization="safetensors", save_ctfidf=True, save_embedding_model=se_model)
-    openai_bertopic_model.save(f"{bertopic_models_dir}/{model_id}/openai", serialization="safetensors", save_ctfidf=True, save_embedding_model=se_model)
+    se_model = f'sentence-transformers/{sentence_transformer_lm}'
+    mmr_bertopic_model.save(f'{BERTOPIC_MODELS_DIR}/{show_key}/{model_id}/mmr', serialization='safetensors', save_ctfidf=True, save_embedding_model=se_model)
+    kb_bertopic_model.save(f'{BERTOPIC_MODELS_DIR}/{show_key}/{model_id}/kb', serialization='safetensors', save_ctfidf=True, save_embedding_model=se_model)
+    openai_bertopic_model.save(f'{BERTOPIC_MODELS_DIR}/{show_key}/{model_id}/openai', serialization='safetensors', save_ctfidf=True, save_embedding_model=se_model)
     # save bertopic_docs_df
-    bertopic_docs_file_name = f'{bertopic_data_dir}/{model_id}.csv'
+    bertopic_docs_file_name = f'{BERTOPIC_DATA_DIR}/{show_key}/{model_id}.csv'
     print(f'writing bertopic_docs_df to file path={bertopic_docs_file_name}')
     bertopic_docs_df.to_csv(bertopic_docs_file_name, sep='\t')
 
@@ -764,13 +622,16 @@ def get_model_response(messages, prompt):
         prompt=prompt,
         messages=messages,
         model='gpt-3.5-turbo-instruct',
-        top_p = 0.5, 
-        max_tokens = 1000,
+        top_p=0.5, 
+        max_tokens=1000,
         stream=True
     )
 
+    # NOTE untested, but this is how sentiment analysis requests were formatted (in Sep 2024 vs May 2024)
+    # response = openai_client.chat.completions.create(
+    #     model="gpt-4o-mini",
+    #     messages=[{"role": "user", "content": prompt}],
+    #     stream=False,
+    # )
+
     return response.choices[0].message['content']
-
-
-if __name__ == '__main__':
-    main()
