+++ app/nlp/narrative_extractor.py
@@ -1,6 +1,6 @@
 from app.show_metadata import ShowKey
 import app.es.es_read_router as esr
-from app.nlp.nlp_metadata import MIN_WORDS_FOR_BERT, MAX_WORDS_FOR_BERT, MIN_SPEAKER_LINES, MIN_SPEAKER_LINE_RATIOS
+from app.nlp.nlp_metadata import MIN_SPEAKER_LINES, MIN_SPEAKER_LINE_RATIOS
 
 
 def extract_narrative_sequences(show_key: ShowKey, episode_key: str) -> list:
