+++ app/page_callbacks/episode_callbacks.py
@@ -3,7 +3,7 @@ from operator import itemgetter
 import os
 import pandas as pd
 
-from app.app_metadata import PATH_TO_SENTIMENT_DATA
+from app.app_metadata import SENTIMENT_DATA_DIR
 import app.data_service.field_flattener as fflat
 import app.es.es_read_router as esr
 import app.fig_builder.plotly_bar as pbar
@@ -190,7 +190,7 @@ def render_episode_sentiment_line_chart(show_key: str, episode_key: str, freeze_
     print(f'in render_episode_sentiment_line_chart, show_key={show_key} episode_key={episode_key} freeze_on={freeze_on} emotion={emotion} speaker={speaker}')
 
    # fetch episode sentiment data and build line chart
-    file_path = f'{PATH_TO_SENTIMENT_DATA}/{show_key}/openai_emo/{show_key}_{episode_key}.csv'
+    file_path = f'{SENTIMENT_DATA_DIR}/{show_key}/openai_emo/{show_key}_{episode_key}.csv'
     if os.path.isfile(file_path):
         df = pd.read_csv(file_path)
         print(f'loading dataframe at file_path={file_path}')
@@ -284,7 +284,7 @@ def render_speaker_frequency_bar_chart(show_key: str, episode_key: str, scale_by
 
     # TODO incorporate episode-level sentiment into es writer workflow; for now it's a quick lookup in episode-level dfs
     emo_limit = 3
-    file_path = f'{PATH_TO_SENTIMENT_DATA}/{show_key}/openai_emo/{show_key}_{episode_key}.csv'
+    file_path = f'{SENTIMENT_DATA_DIR}/{show_key}/openai_emo/{show_key}_{episode_key}.csv'
     if not os.path.isfile(file_path):
         utils.hilite_in_logs(f'No sentiment data found at file_path={file_path}, continuing without it')
         pass
